Validation

+++ src/app/sidebar/docs.service.ts
@@ -13,6 +13,13 @@ export class DocsService {
 
   addNewDoc(doc: Doc) {
     let message = "Added sucessfully";
+    let errorMessage = this.validate(doc);
+   
+    if(errorMessage){
+      alert(errorMessage);
+      return;
+    }
+
     if(this.getDoc(doc.id)){
       this.deleteDoc(doc.id);
       message = "Updated sucesfully"
@@ -24,7 +31,6 @@ export class DocsService {
       doc.id = this.docs[this.docs.length - 1].id + 1;
     }
 
-    
     this.docs.push(doc);
     alert(message);
     this.close();
@@ -49,7 +55,6 @@ export class DocsService {
   }
 
   editDoc(id: number, updatedDoc: Doc) {
-    console.log("editing doc");
     let index = this.getIndex(id);
     this.docs[index] = updatedDoc;
   }
@@ -84,13 +89,43 @@ export class DocsService {
         (<HTMLInputElement>document.getElementById("audience")).value = doc.details.audience;
         (<HTMLInputElement>document.getElementById("overall")).value = doc.details.overall;
         (<HTMLInputElement>document.getElementById("features")).value = doc.details.features;
+        (<HTMLInputElement>document.getElementById("hiddenImage")).value = doc.details.path;
+        (<HTMLInputElement>document.getElementById("img")).src = doc.details.path;
         break;
       case 3:
-        //TODO LATER
+        (<HTMLInputElement>document.getElementById("fileName_1")).value = doc.name;
+        (<HTMLInputElement>document.getElementById("fileImage_1")).src = doc.details.path;
+        (<HTMLInputElement>document.getElementById("hiddenImage_1")).value = doc.details.path;
         break;
       default:
         console.log("An error happened");
         break;
     }
+  }
+  
+  validate(doc:any){
+    let message = "";
+    switch (doc.phase) {
+      case 1:
+        if(!doc.details.projectTitle||!doc.details.endDate||!doc.details.projectObjectives||!doc.details.projectManager||!doc.details.budgetInformation||!doc.details.projectScope||!doc.details.startDate){
+          message = "Please fill all fields";
+        }
+        break;
+      case 2:
+        if(!doc.details.intro||!doc.details.purpose||!doc.details.audience||!doc.details.overall||!doc.details.features||!doc.details.path){
+          message = "Please fill all fields"; 
+        }
+        break;
+      case 3:
+        if(!doc.name || !doc.details.path){
+          message = "Please fill all fields";
+        }
+        break;
+      default:
+        console.log("An error happened");
+        break;
+    }
+    return message;
+
   }
 }
